chore: 각 모델의 구현체를 적절한 위치로 분리
