+++ iOS-MyCarMaster/MyCarMaster/MyCarMaster/Engine/EngineView.swift
@@ -14,14 +14,15 @@ final class EngineView: BasicStepView {
     }
 }
 
-#if canImport(SwiftUI) && DEBUG
-import SwiftUI
-
-struct EngineViewControllerPreviews_Previews: PreviewProvider {
-    static var previews: some View {
-        UIViewControllerPreview {
-            return EngineViewController()
-        }
-    }
+struct Engine {
+    let model: String
+    let name: String
+    let ratio: Int
+    let description: String
+    let fuelMin: Double
+    let fuelMax: Double
+    let power: Int
+    let toque: Double
+    let price: Int
+    var imageURL: URL?
 }
-#endif
