+++ iOS-MyCarMaster/MyCarMaster/MyCarMaster/General/BasicListCell.swift
@@ -201,72 +201,5 @@ extension BasicListCellState {
     }
     init(from wheelDrive: WheelDrive) {
         self.init(model: wheelDrive.model, name: wheelDrive.name, description: wheelDrive.description, ratio: wheelDrive.ratio, price: wheelDrive.price, hasDetailButton: false)
-struct Trim {
-    let model: String
-    let name: String
-    let ratio: Int
-    let description: String
-    var price: Int
-    var imageURL: URL?
-}
-
-struct Engine {
-    let model: String
-    let name: String
-    let ratio: Int
-    let description: String
-    let fuelMin: Double
-    let fuelMax: Double
-    let power: Int
-    let toque: Double
-    let price: Int
-    var imageURL: URL?
-}
-
-#if canImport(SwiftUI) && DEBUG
-import SwiftUI
-
-struct TrimListViewCellPreviews_Previews: PreviewProvider {
-    static let trim = Trim(model: "펠리세이드", name: "Exclusive", ratio: 54, description: "실용적인 기본 기능을 갖춘 베이직 트림", price: 40440000, imageURL: nil)
-
-    static let trimPreview = UIViewPreview {
-        let listViewCell = BasicListCell()
-        let cellState = BasicListCellState(from: trim)
-        listViewCell.configure(with: cellState)
-        listViewCell.select()
-        return listViewCell
-    }
-    static var previews: some View {
-        Group {
-            trimPreview
-                .previewDevice("iPhone SE (3rd generation)")
-            trimPreview
-                .previewDevice("iPhone 14 Pro")
-            trimPreview
-                .previewDevice("iPhone 14 Pro Max")
-        }
     }
 }
-
-struct EngineListViewCellPreviews_Previews: PreviewProvider {
-    static let engine = Engine(model: "펠리세이드", name: "가솔린 3.8", ratio: 54, description: "엔진의 진동이 적어 편안하고 조용한 드라이빙 감성을 제공합니다엔진의 진동이 적어 편안하고 조용한 드라이빙 감성을 제공합니다엔진의 진동이 적어 편안하고 조용한 드라이빙 감성을 제공합니다엔진의 진동이 적어 편안하고 조용한 드라이빙 감성을 제공합니다", fuelMin: 8.0, fuelMax: 9.2, power: 295, toque: 36.2, price: 0, imageURL: nil)
-
-    static let enginePreview = UIViewPreview {
-        let listViewCell = BasicListCell()
-        let cellState = BasicListCellState(from: engine)
-        listViewCell.configure(with: cellState)
-        return listViewCell
-    }
-
-    static var previews: some View {
-        Group {
-            enginePreview
-                .previewDevice("iPhone SE (3rd generation)")
-            enginePreview
-                .previewDevice("iPhone 14 Pro")
-            enginePreview
-                .previewDevice("iPhone 14 Pro Max")
-        }
-    }
-}
-#endif
